requirements.txt
adtk==0.6.2
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessings==1.7
bokeh==3.1.1
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
codecov==2.1.13
colorama==0.4.6
colorcet==3.0.1
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.1.3
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.1
distlib==0.3.7
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
Fiona==1.9.4.post1
fonttools==4.41.1
frozenlist==1.4.0
geoip2==4.7.0
geopandas==0.13.2
holoviews==1.17.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter-client==8.3.0
jupyter-core==5.3.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
maxminddb==2.4.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.4
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
panel==1.2.1
param==1.13.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
powerlaw==1.5
pprintpp==0.4.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-radix==0.10.0
pyasn==1.6.1
pycountry==22.3.5
pycountry-convert==0.7.2
pyct==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.5.0
pyproject-api==1.5.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
pytz==2023.3
pyviz-comms==2.3.2
PyYAML==6.0.1
pyzmq==25.1.0
repoze.lru==0.7
requests==2.31.0
rootpath==0.1.1
scikit-learn==1.3.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.2
tox==4.6.4
tqdm==4.66.0
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
virtualenv==20.24.2
wcwidth==0.2.6
webencodings==0.5.1
wget==3.2
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.16.2
